requirements.txt
python-telegram-bot[rate-limiter]==21.*
apscheduler
httpx
